fix: systematic PHPStan level 8 compliance improvements (50+ errors fixed)
Reduced PHPStan errors from 214 to 164 through systematic easy-fix approach:

## HTTP Client Package Improvements
- Fixed impossible comparison in PoolCurlResponseAdapter (string|null vs false)
- Added proper null safety checks for curl_multi_getcontent() results
- Added getter methods for "unused" properties in CurlPool (events)
- Removed dead code methods in response adapters:
  * Guzzle: statusCode(), headers(), isStreamed()
  * Laravel: statusCode(), headers(), isStreamed()
  * ExtHttp: headers() method
  * Mock: stream() method
- Fixed strict type comparisons

## Laravel Eloquent Model Type Safety
- Added proper type annotations for Eloquent model properties:
  * Issue: $fillable and $casts arrays
  * Comment: $fillable and $casts arrays
  * Dependency: $fillable and $casts arrays
- Pattern: @var array<string> for $fillable, @var array<string, string> for $casts

## Logging Package Improvements
- Fixed callable return type signatures in LaravelLoggingFactory
- Added detailed @return callable(Event): void annotations
- Imported missing Event class for proper typing
- Enhanced MonologChannelWriter type safety
- Improved LazyEnricher closure type annotations

## Console Command Fixes
- Removed redundant $defaultName properties in Beads commands
- Fixed property conflicts where constructor already sets command names
- Applied to 9+ command classes in auxiliary package

## Code Quality Fixes
- Fixed critical resolveToolKey() code duplication bug in PartialInferenceResponse
- Removed redundant type casting in Usage.php
- Fixed test class naming conflicts (TestEvent -> specific names)
- Added proper getter methods instead of deleting "unused" properties

## Monorepo Integration
- Added logging package to phpunit.xml test paths
- Added logging package to PHPStan configuration
- Updated packages.json for GitHub workflow automation
- Integrated logging package into CI/CD pipeline

📈 Error Reduction: 214 → 164 errors (23% improvement)
🎯 Focus: Easy mechanical fixes, proven patterns, systematic approach

+++ packages/auxiliary/src/Beads/Infrastructure/Execution/SandboxCommandExecutor.php
@@ -62,7 +62,7 @@ final class SandboxCommandExecutor implements CommandExecutor
                 }
 
                 // Exponential backoff: 100ms, 200ms, 400ms, etc.
-                $exponent = (int) (2 ** ($attempt - 1));
+                $exponent = 2 ** ($attempt - 1);
                 $backoffMs = 100 * $exponent;
                 usleep($backoffMs * 1000);
             }
